DDF-3551 Refactored delivery information gathering to get user preferences once
I refactored the code to retrieve delivery information so that instead of retrieving the user's preferences for each delivery method used, it is retrieved once for all delivery methods on a schedule.

DDF-3551 Renamed username related variables to indicate user ID

Previously, the username was passed into the schedule data and used to identify user preferences; variables were named to indicate such. However, another change to DDF changed both the identifier used to get the user preferences and the identifier put inside the schedule data; as such, the function of this code is not affected, but I changed the variable names to clarify the use of the user ID instead of a username.

DDF-3551 Added and improved documentation, especially in QueryDeliveryService

The QueryDeliveryService interface is now documented for use in creating other delivery services. In addition, some documentation was added and improved for Fallible Ignite objects in the QuerySchedulingPostIngestPlugin.

DDF-3551 Removed a TODO comment that was no longer relevant

We were using "userId" as the key when it should have been "username" for a time, but then code changed and we are actually using the user ID now, so the change no longer needs to be made.

DDF-3551 Improved Fallible.forEach and improved process error handling

The process function implementations used an error collection and handling that was harder to read, less efficient, and more redundant than the Fallible.forEach, which was not implemented at the time. Additionally, Fallible.forEach can now take in multiple mapper functions and collect errors from all of them. Finally, Fallible.forEach is now documented.

DDF-3551 Ignite handling fixes and improvements

With these changes, the running queries cache and scheduler from Ignite will be checked and retrieved from a (hopefully) running Ignite instance each time a job is scheduled, run, or cancelled. This will make the code cleaner and more robust in the context of Ignite instance shutdowns.

DDF-3551 Scheduling and Fallible.forEach bug fixes

Scheduling behavior was corrected and stabilized with some added end date checks.

A RuntimeException from reusing a finished stream was fixed in Fallible.forEach.